% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_dyn_df}
\alias{create_dyn_df}
\title{create_din_df}
\usage{
create_dyn_df(data_list, col_names, round_digits)
}
\arguments{
\item{data_list}{A list of data frames or vectors to be combined into a single data frame.}

\item{col_names}{A character vector of column names to assign to the resulting data frame.}

\item{round_digits}{Numeric. The number of decimal places to round numeric values in the data to.}
}
\value{
A data frame with the combined data and rounded numeric values.
}
\description{
Creates a dynamic data frame by combining a list of data and rounding numeric values to a specified number of digits.
This function takes a list of data, rounds the numeric values, and combines them into a single data frame with custom column names.
}
