% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_df_example.R
\docType{data}
\name{forest_df_example}
\alias{forest_df_example}
\title{Data set about forest dynamics}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{plot}{The plot identification number in the field.}
 \item{n}{The identification number of the trees.}
 \item{spp}{The scientific name of the species}
 \item{DBH_1}{Numeric. The diameter at breast height (DBH) measurement on the first inventory.}
 \item{DBH_2}{Numeric. The diameter at breast height (DBH) measurement on the second inventory.}
 }
}
\usage{
data(forest_df_example)
}
\description{
Data set about forest dynamics
}
\examples{
head(forest_df_example)


}
\references{
This dataset was created by Pedro Higuchi, based on his filed research.
}
\keyword{dataset}
