% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_creation.R
\name{save_dyn_files}
\alias{save_dyn_files}
\title{save_dyn_files}
\usage{
save_dyn_files(dynamics)
}
\arguments{
\item{dynamics}{A list containing four data frames with different forest dynamics data:
1) number of plots (`n_plot`),
2) number of species (`n_species`),
3) basal area by species (`basal_area_species`),
4) basal area by plots (`basal_area_plot`).}
}
\value{
NULL If the user does not select a folder, the function will return NULL and no files will be saved.
             Otherwise, the function saves four CSV files and returns a confirmation message.
}
\description{
This function allows the user to save dynamic forest data to CSV files. It prompts the user
to select a directory via a graphical interface (using the `tcltk` package), and then saves
several data frames containing forest dynamics information to predefined file names in that directory.
}
